# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        },
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/recipient_lists": {
            "get": {
                "tags": [
                    "Recipient Lists"
                ],
                "summary": "Get a list of recipient lists",
                "operationId": "a95b3f9bb06c6e18f65a0292e2f6d2b8",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipient Lists"
                ],
                "summary": "Store a new recipient list",
                "operationId": "07290f82ff8d20bc439f4a116850db16",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "List Name"
                                    },
                                    "csv_file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recipient list created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recipient_lists/{id}": {
            "get": {
                "tags": [
                    "Recipient Lists"
                ],
                "summary": "Get a recipient list",
                "operationId": "b903cd9afcde64dea17c9d73ca6a0ba2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recipient List ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "recipientList": {
                                            "type": "object"
                                        },
                                        "contacts": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Recipient Lists"
                ],
                "summary": "Update a recipient list",
                "operationId": "a78d4efe5ee7080c814e2bf25f1b3c63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recipient List ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated List Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recipient list updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "recipientList": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Recipient Lists"
                ],
                "summary": "Delete a recipient list",
                "operationId": "e33e7098223935abb02e7454ce761b9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recipient List ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipient list deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "List deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}